Améliorations mineures, testScenario updated, formattage, création d'exceptions

+++ src/Autre/Chemin.java
@@ -20,7 +20,7 @@ public class Chemin {
         return chemin;
     }
 
-    public void addElement(Case caseSuiv, int date) {
+    public void addElement(Case caseSuiv, long date) {
         this.chemin.add(new AssociationTempsCase(caseSuiv, date));
     }
 
@@ -39,6 +39,9 @@ public class Chemin {
      * 
      * @param simulateur : Simulateur permettant l'ajout d'évènements
      * @param robot      : Robot concerné par le déplacement sur le chemin
+     * 
+     * @exception IllegalCheminRobotException Le chemin n'est pas traversable par le
+     *                                        robot
      */
     public void creerEvenements(Simulateur simulateur, Robot robot) throws IllegalCheminRobotException {
         for (AssociationTempsCase tc : this.chemin) {
@@ -62,9 +65,9 @@ public class Chemin {
 class AssociationTempsCase {
 
     private Case _case;
-    private int t;
+    private long t;
 
-    public AssociationTempsCase(Case _case, int t) {
+    public AssociationTempsCase(Case _case, long t) {
         this._case = _case;
         this.t = t;
     }
@@ -73,7 +76,7 @@ class AssociationTempsCase {
         return _case;
     }
 
-    public int getT() {
+    public long getT() {
         return t;
     }
